llm_basics/requirements.txt
openai==1.3.0
python-dotenv==1.0.0
tiktoken==0.5.1